PO-3 - added unit test for UserDtoConverterImpl, changed user_role id type to TINYINT and reordered UserRoleEnum according to order in DB.

+++ core/src/main/java/priceobserver/data/userrole/UserRole.java
@@ -13,16 +13,16 @@ public class UserRole {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Long id;
+    private Byte id;
 
     @NotNull
     private String name;
 
-    public Long getId() {
+    public Byte getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Byte id) {
         this.id = id;
     }
